* HTTPS via gnutls * Many bug fixes and enhancements

+++ hash.h
@@ -30,8 +30,8 @@
 #include "mget.h"
 
 int
-   hash_file_fd(const char *type, int fd, char *digest_hex, size_t digest_hex_size, off_t offset, size_t length),
-   hash_file_offset(const char *type, const char *fname, char *digest_hex, size_t digest_hex_size, off_t offset, size_t length),
+   hash_file_fd(const char *type, int fd, char *digest_hex, size_t digest_hex_size, off_t offset, off_t length),
+   hash_file_offset(const char *type, const char *fname, char *digest_hex, size_t digest_hex_size, off_t offset, off_t length),
    hash_file(const char *type, const char *fname, char *digest_hex, size_t digest_hex_size);
